🔢 FIX: Support decimal exam scores (e.g., 97.50, 97.75)
Changes:
- monthly-exams.js: Use parseFloat instead of parseInt for exam scores
- honor-system.js: Display scores with 2 decimal places (.toFixed(2))
- Ensure accurate sorting by precise decimal scores
- Both table view and PDF export now show decimal precision

Benefits:
- Scores like 97.50 and 97.75 are now preserved
- More accurate ranking based on precise scores
- Better differentiation between top students

+++ js/monthly-exams.js
@@ -170,7 +170,7 @@ window.saveMonthlyExam = async function() {
   const teacherId = document.getElementById('examTeacherSelect')?.value;
   const studentId = document.getElementById('examStudentSelect')?.value;
   const scope = document.getElementById('examScope')?.value.trim();
-  const score = parseInt(document.getElementById('examScore')?.value);
+  const score = parseFloat(document.getElementById('examScore')?.value);
   const notes = document.getElementById('examNotes')?.value.trim();
   
   const messageBox = document.getElementById('examFormMessage');
@@ -766,7 +766,7 @@ window.closeEditExam = function() {
  */
 window.saveEditExam = async function(examId) {
   const scope = document.getElementById('editExamScope')?.value.trim();
-  const score = parseInt(document.getElementById('editExamScore')?.value);
+  const score = parseFloat(document.getElementById('editExamScore')?.value);
   const notes = document.getElementById('editExamNotes')?.value.trim();
   const messageDiv = document.getElementById('editFormMessage');
